Move code to gui and recover volcano test
