+++ alphastats/gui/utils/analysis.py
@@ -2,13 +2,16 @@
 
 from abc import ABC, abstractmethod
 from collections import defaultdict
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import Any, Dict, Literal, Optional, Tuple, Union
 
 import pandas as pd
 import streamlit as st
+from plotly.graph_objs._figure import Figure
 
 from alphastats.dataset.dataset import DataSet
 from alphastats.dataset.keys import Cols, ConstantsClass
+from alphastats.gui.utils.ui_helper import StateKeys
+from alphastats.pl.volcano import _plot_volcano, prepare_result_df
 from alphastats.plots.plot_utils import PlotlyObject
 from alphastats.plots.volcano_plot import VolcanoPlot
 
@@ -514,3 +517,265 @@ ANALYSIS_OPTIONS = {
     StatisticOptions.ANOVA: AnovaAnalysis,
     StatisticOptions.ANCOVA: AncovaAnalysis,
 }
+
+
+class ResultObject(ABC):
+    def __init__(
+        self,
+        dataframe: pd.DataFrame,
+        plottable: bool,
+        preprocessing: Dict,
+        method: Dict,
+    ):
+        if dataframe is None:
+            raise ValueError("Either dataframe or plot must be provided")
+
+        self.dataframe = dataframe
+        self.annotated_dataframe = dataframe
+        self.plottable = plottable
+        self.plot = None
+        self.preprocessing = preprocessing
+        self.method = method
+        self.data_annotation_options = {}
+        self.display_options = {}
+        self.display_selection = "Plot" if self.plottable else "Dataframe"
+
+    def _apply_data_annotation_options(self):
+        """Function to get all significance options for the analysis object"""
+        self.data_annotation_options = self._get_data_annotation_options()
+        self.annotated_dataframe = self._update_data_annotation(
+            **self.data_annotation_options
+        )
+
+    @abstractmethod
+    def _get_data_annotation_options(self) -> Dict:
+        pass
+
+    @abstractmethod
+    def _update_data_annotation(self, **kwargs) -> pd.DataFrame:
+        pass
+
+    def _apply_display_options(self):
+        """Funciton to get all display options for the analysis object"""
+        if self.plottable is False:
+            display_selection_options = ["Raw dataframe", "Annotated dataframe"]
+        else:
+            display_selection_options = ["Plot", "Raw dataframe", "Annotated dataframe"]
+        display_selection = st.radio(
+            "Select display",
+            display_selection_options,
+            index=display_selection_options.index(
+                st.session_state.get("TMP_display_selection", self.display_selection)
+            ),
+            key="TMP_display_selection",
+        )
+        self.display_selection = display_selection
+        if display_selection == "Plot":
+            self.display_options = self._get_plot_options()
+            self.plot = self._update_plot(**self.display_options)
+
+    @abstractmethod
+    def _get_plot_options(self) -> Dict:
+        """Function to get all plotting related options for the analysis object"""
+        pass
+
+    @abstractmethod
+    def _update_plot(self, **kwargs) -> Figure:
+        """Function to update the display of the analysis object"""
+        pass
+
+    def _display_object(self):
+        """Function to display the analysis object"""
+        if self.display_selection == "Plot":
+            st.plotly_chart(self.plot.update())
+        elif self.display_selection == "Raw dataframe":
+            st.dataframe(self.dataframe)
+        elif self.display_selection == "Annotated dataframe":
+            st.dataframe(self.annotated_dataframe)
+
+    def display_object(
+        self,
+        display_column: st.container,
+        data_annotation_editable: bool = False,
+        display_editable: bool = False,
+        widget_column: Optional[st.container] = None,
+    ):
+        """Function to display the object.
+        The function will display the object in the display column and the options in the widget column.
+        The boolean flags are intended for controlling behaviour in different sections of the application.
+
+        Args:
+            display_column (st.container): The container to display the object.
+            data_annotation_editable (bool, optional): If the data_annotation options are editable. Defaults to False.
+            display_editable (bool, optional): If the display options are editable. Defaults to False.
+            widget_column (Optional[st.container], optional): The container to display the widgets. Defaults to None.
+
+        Raises:
+            ValueError: If the widget column container is not provided.
+        """
+        if data_annotation_editable or display_editable:
+            if widget_column is None:
+                raise ValueError("Widget column container must be provided")
+            with widget_column:
+                if data_annotation_editable:
+                    self._apply_data_annotation_options()
+                if display_editable:
+                    self._apply_display_options()
+                else:
+                    self.plot = self._update_plot(**self.display_options)
+        with display_column:
+            self._display_object()
+
+    def get_standard_layout_options(self):
+        return {
+            "height": st.number_input(
+                "Height",
+                200,
+                1000,
+                st.session_state.get(
+                    "TMP_height", self.display_options.get("height", 500)
+                ),
+                10,
+                key="TMP_height",
+            ),
+            "width": st.number_input(
+                "Width",
+                200,
+                1000,
+                st.session_state.get(
+                    "TMP_width", self.display_options.get("width", 500)
+                ),
+                10,
+                key="TMP_width",
+            ),
+            "showlegend": st.checkbox(
+                "Show legend",
+                st.session_state.get(
+                    "TMP_showlegend", self.display_options.get("showlegend", 500)
+                ),
+                key="TMP_showlegend",
+            ),
+        }
+
+
+class DifferentialExpressionTwoGroupsResult(ResultObject):
+    def __init__(
+        self,
+        dataframe: pd.DataFrame,
+        preprocessing: Dict,
+        method: Dict,
+    ):
+        super().__init__(
+            dataframe, plottable=True, preprocessing=preprocessing, method=method
+        )
+        self.log2name = ""
+
+    def _get_data_annotation_options(self) -> Dict:
+        return {
+            "qvalue_cutoff": st.number_input(
+                "Q-value cutoff",
+                0.0,
+                1.0,
+                st.session_state.get(
+                    "TMP_qvalue_cutoff",
+                    self.data_annotation_options.get("qvalue_cutoff", 0.05),
+                ),
+                0.01,
+                format="%.2f",
+                key="TMP_qvalue_cutoff",
+            ),
+            "log2fc_cutoff": st.number_input(
+                "Log2FC cutoff",
+                0.0,
+                10.0,
+                st.session_state.get(
+                    "TMP_log2fc_cutoff",
+                    self.data_annotation_options.get("log2fc_cutoff", 1.0),
+                ),
+                0.1,
+                format="%.1f",
+                key="TMP_log2fc_cutoff",
+            ),
+            "flip_xaxis": st.checkbox(
+                "Flip groups",
+                st.session_state.get(
+                    "TMP_flip_xaxis",
+                    self.data_annotation_options.get("flip_xaxis", False),
+                ),
+                key="TMP_flip_xaxis",
+            ),
+        }
+
+    def _update_data_annotation(
+        self,
+        qvalue_cutoff: float,
+        log2fc_cutoff: float,
+        flip_xaxis: bool,
+    ) -> pd.DataFrame:
+        formatted_df, log2name = prepare_result_df(
+            statistics_results=self.dataframe,
+            feature_to_repr_map=st.session_state[StateKeys.DATASET].feature_to_repr_map,
+            group1=self.method["group1"],
+            group2=self.method["group2"],
+            qvalue_cutoff=qvalue_cutoff,
+            log2fc_cutoff=log2fc_cutoff,
+            flip_xaxis=flip_xaxis,
+        )
+        self.log2name = log2name
+        return formatted_df
+
+    def _get_plot_options(self) -> Dict:
+        with st.expander("Display options"):
+            renderer_options = ["webgl", "svg"]
+            return {
+                **{
+                    "drawlines": st.checkbox(
+                        "Draw significance and fold change lines",
+                        st.session_state.get(
+                            "TMP_drawlines", self.display_options.get("drawlines", True)
+                        ),
+                        key="TMP_drawlines",
+                    ),
+                    "label_significant": st.checkbox(
+                        "Label significant points",
+                        st.session_state.get(
+                            "TMP_label_significant",
+                            self.display_options.get("label_significant", True),
+                        ),
+                        key="TMP_label_significant",
+                    ),
+                    "renderer": st.radio(
+                        "Renderer",
+                        renderer_options,
+                        index=renderer_options.index(
+                            st.session_state.get(
+                                "TMP_renderer",
+                                self.display_options.get("renderer", "webgl"),
+                            )
+                        ),
+                        key="TMP_renderer",
+                    ),
+                },
+                **self.get_standard_layout_options(),
+            }
+
+    def _update_plot(
+        self,
+        drawlines: bool,
+        label_significant: bool,
+        renderer: Literal["webgl", "svg"],
+        **kwargs,
+    ) -> Figure:
+        return _plot_volcano(
+            df_plot=self.annotated_dataframe,
+            log2name=self.log2name,
+            group1=self.method["group1"],
+            group2=self.method["group2"],
+            qvalue_cutoff=self.data_annotation_options["qvalue_cutoff"],
+            log2fc_cutoff=self.data_annotation_options["log2fc_cutoff"],
+            flip_xaxis=self.data_annotation_options["flip_xaxis"],
+            drawlines=drawlines,
+            label_significant=label_significant,
+            renderer=renderer,
+            **kwargs,
+        )
