+++ alphastats/pl/volcano.py
@@ -3,11 +3,8 @@ from typing import Dict, Literal, Tuple, Union
 import numpy as np
 import pandas as pd
 import plotly.express as px
-import streamlit as st
 from plotly.graph_objs._figure import Figure
 
-from alphastats.gui.utils.ui_helper import StateKeys
-from alphastats.pl.display_utils import ResultObject
 from alphastats.tl.differential_expression_analysis import (
     DeaColumns,
     DifferentialExpressionAnalysis,
@@ -297,126 +294,3 @@ def prepare_result_df(
     result_df["label"] = result_df.index.map(feature_to_repr_map)
 
     return result_df, log2name
-
-
-class DifferentialExpressionTwoGroupsResult(ResultObject):
-    def __init__(
-        self,
-        dataframe: pd.DataFrame,
-        preprocessing: Dict,
-        method: Dict,
-    ):
-        super().__init__(
-            dataframe, plottable=True, preprocessing=preprocessing, method=method
-        )
-        self.log2name = ""
-
-    def _get_data_annotation_options(self) -> Dict:
-        return {
-            "qvalue_cutoff": st.number_input(
-                "Q-value cutoff",
-                0.0,
-                1.0,
-                st.session_state.get(
-                    "TMP_qvalue_cutoff",
-                    self.data_annotation_options.get("qvalue_cutoff", 0.05),
-                ),
-                0.01,
-                format="%.2f",
-                key="TMP_qvalue_cutoff",
-            ),
-            "log2fc_cutoff": st.number_input(
-                "Log2FC cutoff",
-                0.0,
-                10.0,
-                st.session_state.get(
-                    "TMP_log2fc_cutoff",
-                    self.data_annotation_options.get("log2fc_cutoff", 1.0),
-                ),
-                0.1,
-                format="%.1f",
-                key="TMP_log2fc_cutoff",
-            ),
-            "flip_xaxis": st.checkbox(
-                "Flip groups",
-                st.session_state.get(
-                    "TMP_flip_xaxis",
-                    self.data_annotation_options.get("flip_xaxis", False),
-                ),
-                key="TMP_flip_xaxis",
-            ),
-        }
-
-    def _update_data_annotation(
-        self,
-        qvalue_cutoff: float,
-        log2fc_cutoff: float,
-        flip_xaxis: bool,
-    ) -> pd.DataFrame:
-        formatted_df, log2name = prepare_result_df(
-            statistics_results=self.dataframe,
-            feature_to_repr_map=st.session_state[StateKeys.DATASET].feature_to_repr_map,
-            group1=self.method["group1"],
-            group2=self.method["group2"],
-            qvalue_cutoff=qvalue_cutoff,
-            log2fc_cutoff=log2fc_cutoff,
-            flip_xaxis=flip_xaxis,
-        )
-        self.log2name = log2name
-        return formatted_df
-
-    def _get_plot_options(self) -> Dict:
-        with st.expander("Display options"):
-            renderer_options = ["webgl", "svg"]
-            return {
-                **{
-                    "drawlines": st.checkbox(
-                        "Draw significance and fold change lines",
-                        st.session_state.get(
-                            "TMP_drawlines", self.display_options.get("drawlines", True)
-                        ),
-                        key="TMP_drawlines",
-                    ),
-                    "label_significant": st.checkbox(
-                        "Label significant points",
-                        st.session_state.get(
-                            "TMP_label_significant",
-                            self.display_options.get("label_significant", True),
-                        ),
-                        key="TMP_label_significant",
-                    ),
-                    "renderer": st.radio(
-                        "Renderer",
-                        renderer_options,
-                        index=renderer_options.index(
-                            st.session_state.get(
-                                "TMP_renderer",
-                                self.display_options.get("renderer", "webgl"),
-                            )
-                        ),
-                        key="TMP_renderer",
-                    ),
-                },
-                **self.get_standard_layout_options(),
-            }
-
-    def _update_plot(
-        self,
-        drawlines: bool,
-        label_significant: bool,
-        renderer: Literal["webgl", "svg"],
-        **kwargs,
-    ) -> Figure:
-        return _plot_volcano(
-            df_plot=self.annotated_dataframe,
-            log2name=self.log2name,
-            group1=self.method["group1"],
-            group2=self.method["group2"],
-            qvalue_cutoff=self.data_annotation_options["qvalue_cutoff"],
-            log2fc_cutoff=self.data_annotation_options["log2fc_cutoff"],
-            flip_xaxis=self.data_annotation_options["flip_xaxis"],
-            drawlines=drawlines,
-            label_significant=label_significant,
-            renderer=renderer,
-            **kwargs,
-        )
